Create Code_Rmd_Format
This code was written in the R markdown file in R to make a PDF doc containing the detail report of the project. Rmd was converted to md which was then converted to pdf using the Pander package in R. The minimum requirements for converting a md file to pdf is Pandoc, Miktex.